+++ styles/components/setting.ts
@@ -82,8 +82,6 @@ const button = css`
   .admin-setting-content {
     .ui.primary.button {
       ${tinyStyle}
-      padding: 5px 16px;
-      line-height: 22px;
       box-shadow: none;
       &:hover {
         ${tinyHoverStyle}
@@ -96,10 +94,6 @@ const button = css`
         border-color: ${themeVars.github.button.primary.borderColor.hover};
         color: ${themeVars.github.button.primary.fgColor.rest};
       }
-    }
-    .ui.red.button {
-      padding: 5px 16px;
-      line-height: 22px;
     }
   }
 `;