+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ### 🐞 Fix
 
 - Fix the height issue of the global small button.
+- Fix the height issue of the administrator interface button.
+- Fix the height issue of the code copy button.